+++ frontend/src/Components/FilterSideBar.js
@@ -1,0 +1,9 @@
+import { filterSideBarClassName } from "../commonStyling"
+
+
+function FilterSideBar(){
+
+    return (<div className={filterSideBarClassName}>Placeholder for filters</div>)
+}
+
+export default  FilterSideBar