+++ frontend/src/commonStyling.js
@@ -1,0 +1,13 @@
+const color = "gray"
+
+const formStyle = `flex flex-col bg-${color}-100 p-4 rounded-md items-center`
+const formElementStyle = `w-80 px-2 py-2 m-2 border rounded-md `
+const buttonClass = `w-40 py-2 m-2 bg-${color}-400 hover:bg-${color}-500 text-white py-2 px-4 rounded-xl`
+const buttonClassGo = `bg-${color}-400 hover:bg-${color}-500 text-white py-2 px-4 rounded-xl`
+const headerClass = `bg-${color}-100 border-b p-5 flex justify-between items-center"`
+const filterSideBarClassName = `"bg-${color}-100 w-1/5 h-screen"`
+
+
+
+
+export {formElementStyle, formStyle, buttonClass, buttonClassGo, headerClass, filterSideBarClassName}