+++ frontend/src/index.js
@@ -6,7 +6,6 @@ import reportWebVitals from './reportWebVitals';
 import Header from './Components/Header';
 
 
-
 const root = ReactDOM.createRoot(document.getElementById('root'));
 
 root.render(<App />);